// Gestion-Pieces--front-end-main - Copie/src/app/models/user.model.ts
export class User {
  id?: number;
  name?: string;
  mobileNumber?: string;
  email?: string;
  address?: string;
  city?: string;
  state?: string;
  pincode?: string;
  password?: string;
  profileImage?: string;
  role?: string;
  isEnable?: boolean;
  accountNonLocked?: boolean;
  failedAttempt?: number;
  lockTime?: Date;
  resetToken?: string;
  }
  